v0.2.0 feat (LambdaInterface, SQSInterface) - checks for destroy() feat (architecture_utils) - cleanup_protocol feat (bus_reduce, catalog_build) - destroy previous lambdas

+++ scito_count/LambdaInterface.py
@@ -37,4 +37,5 @@ class LambdaInterface(object):
         return response
 
     def destroy(self):
-        self.aws_lambda.delete_function(FunctionName=self.lambda_name)
+        if self.function_exists():
+            self.aws_lambda.delete_function(FunctionName=self.lambda_name)
